+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+- Add syntax diagram(index.php, index.html).
 
 ## [0.0.1] - 2023-03-30 [YANKED]
 ### Security